Refactor development Docker Image to ease Usage
To ease the usage of our docker file and improve CI performance, we
introduce some changes to our current Dockerfile. The main changes are

- Providing a top-level docker-compose.yaml file which is used by both
  the CI and local developers to build and run the development docker
  image.
- Provide a lightweight solution for custom UID/GIDs without the need of
  embedding the UID/GID into the image.
- Cleanup installation paths of tooling to be more consistent with
  general rules like the FHS.
- Update CI workflows to also use the top-level docker-compose.yaml and
  cache the docker image for performance reasons.

With the introduction of a pre-built docker image this commits prepares
a docker-compose file for easy local development. Also the CI workflows
now use docker compose to run the different commands. This has the
advantage, that it is easy for a developer to reproduce and we have a
central place to define the image URL to use.