Disable Module Support in CMake
We currently do not use modules in the project and CMake requires for
Clang a clang-scan-deps utility which is currently not shipped with our
ARM toolchain. Therefore, we disable module support for now, in case we
need it later we need to adapt/extend our toolchains such that they
include all necessary tools.