Install Python Dependencies from requirements.lock
To ensure we have fixed and reproducible Python packages installed in
our Docker image, we provide a requirements.lock file created by
pip-compile from all requirements we want to include. This lock file
then contains already resolved dependencies and specifies for each
package an already frozen version which we install.

Since the frozen packages in the lock file potentially conflict with
system packages installed by Ubunut, we do not install this globally
using the --break-system-packages flag (actually this also fails for us
right now) but we provide an already activated venv. This allows us to
be independent of the system packages installed by Ubuntu and really
ensure that we use the packages from our requirements.lock file.