Remove cmake Cache in Github Actions
Until now we cache the cmake build directory in the GitHub actions cache
and restore it in case we see a cache hit. This introduces several
problems:

- The cache key is a manually selected set of files, in case anything
  misses here we incorrectly take the cache.
- We are not confident that an incremental build will work in all cases.
  This is currently happening due to the restore-keys. Especially, since
  ninja uses timestamps for detecting changes which can be wrong.
- The build time is still not too high, so building again might be
  acceptable right now.

In case we want to have a cache in the future we need to carefully think
on how to properly integrate it. Instead of caching the cmake build
directory we could think of introducing a compiler cache like sccache
which also uses hashes/content of files and not timestamps or even
switch to a different build system like bazel which is better prepared
for caching. Also, we need to see how it can be integrated into the
GitHub actions infrastructure.